+++ Beelzebub/Beelzebub/EntityExtensions.cs
@@ -1,0 +1,51 @@
+using ProjectM;
+using ProjectM.Network;
+using Stunlock.Core;
+using Unity.Entities;
+
+namespace Beelzebub;
+
+internal static class EntityExtensions
+{
+    public static bool Exists(this Entity entity) =>
+        entity != Entity.Null && Core.EntityManager.Exists(entity);
+
+    public static bool TryGetComponent<T>(this Entity entity, out T component) where T : unmanaged
+    {
+        if (!entity.Exists() || !Core.EntityManager.HasComponent<T>(entity))
+        {
+            component = default;
+            return false;
+        }
+        component = Core.EntityManager.GetComponentData<T>(entity);
+        return true;
+    }
+
+    public static bool Has<T>(this Entity entity) =>
+        entity.Exists() && Core.EntityManager.HasComponent<T>(entity);
+
+    public static bool IsPlayer(this Entity entity) =>
+        entity.Has<PlayerCharacter>();
+
+    public static bool TryGetPlayer(this Entity entity, out Entity playerCharacter)
+    {
+        if (entity.IsPlayer()) { playerCharacter = entity; return true; }
+        playerCharacter = Entity.Null;
+        return false;
+    }
+
+    public static ulong GetSteamId(this Entity playerCharacter)
+    {
+        if (playerCharacter.TryGetComponent<PlayerCharacter>(out var pc)
+            && pc.UserEntity.TryGetComponent<User>(out var user))
+        {
+            return user.PlatformId;
+        }
+        return 0;
+    }
+
+    public static PrefabGUID GetPrefabGuid(this Entity entity) =>
+        entity.TryGetComponent<PrefabGUID>(out var g) ? g : default;
+
+    public static string GetPrefabName(this PrefabGUID guid) => $"PrefabGuid({guid._Value})";
+}